Rename/#6/컴포넌트명 변경
